Trait dusk_bytes::Read

pub trait Read {
    fn capacity(&self) -> usize;

    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Error>;
}

Implementors of the Read trait are called ‘readers’.

Readers are defined by one required method, [read()]. Each call to [read()] will attempt to pull bytes from this source into a provided buffer.

Required methods

fn capacity(&self) -> usize

Returns the number of elements the Reader can hold.

fn read(&mut self, buf: &mut [u8]) -> Result<usize, Error>

Pull some bytes from this source into the specified buffer, returning how many bytes were read.